+++ packages/launcher/src/App.tsx
@@ -12,7 +12,7 @@ import { saTheme } from '@sea-launcher/style';
 import { CommandBar } from './CommandBar';
 import { MainPanel } from './views/MainPanel';
 
-import { Hook, SEAHookEmitter as HookDispatcher } from 'sea-core';
+import { Hook, SEAHookEmitter } from 'sea-core';
 import { Manager as BattleManager, autoStrategy } from 'sea-core/battle';
 import { EventBus, SocketEventEmitter } from 'sea-core/emitter';
 
@@ -57,29 +57,29 @@ export default function App() {
 
         document.body.addEventListener('keydown', handleShortCut);
 
-        const SEAHookDispatcher = eventBus.proxy(HookDispatcher);
-        const SEASocketListener = eventBus.proxy(SocketEventEmitter);
+        const hookEmitter = eventBus.delegate(SEAHookEmitter);
+        const socketEmitter = eventBus.delegate(SocketEventEmitter);
 
-        SEAHookDispatcher.on(Hook.Battle.battleStart, () => {
+        hookEmitter.on(Hook.Battle.battleStart, () => {
             toggleFighting(true);
         });
 
-        SEAHookDispatcher.on(Hook.Battle.battleStart, handleBattleRoundEnd);
-        SEAHookDispatcher.on(Hook.Battle.roundEnd, handleBattleRoundEnd);
-        SEAHookDispatcher.on(Hook.Battle.battleEnd, () => {
+        hookEmitter.on(Hook.Battle.battleStart, handleBattleRoundEnd);
+        hookEmitter.on(Hook.Battle.roundEnd, handleBattleRoundEnd);
+        hookEmitter.on(Hook.Battle.battleEnd, () => {
             toggleFighting(false);
         });
 
-        SEAHookDispatcher.on(Hook.Battle.battleStart, () => {
+        hookEmitter.on(Hook.Battle.battleStart, () => {
             Logger.BattleManager.info(`检测到对战开始`);
         });
 
-        SEAHookDispatcher.on(Hook.Battle.battleEnd, () => {
+        hookEmitter.on(Hook.Battle.battleEnd, () => {
             const win = Boolean(FightManager.isWin);
             Logger.BattleManager.info(`检测到对战结束 对战胜利: ${win}`);
         });
 
-        SEAHookDispatcher.on(Hook.Award.receive, (data) => {
+        hookEmitter.on(Hook.Award.receive, (data) => {
             Logger.AwardManager.info(`获得物品:`);
             const logStr = Array.isArray(data.items)
                 ? data.items.map((v) => `${ItemXMLInfo.getName(v.id)} ${v.count}`)
@@ -87,15 +87,15 @@ export default function App() {
             logStr && Logger.AwardManager.info(logStr.join('\r\n'));
         });
 
-        SEAHookDispatcher.on(Hook.Module.loadScript, (name) => {
+        hookEmitter.on(Hook.Module.loadScript, (name) => {
             Logger.ModuleManger.info(`检测到新模块加载: ${name}`);
         });
 
-        SEAHookDispatcher.on(Hook.Module.openMainPanel, ({ module, panel }) => {
+        hookEmitter.on(Hook.Module.openMainPanel, ({ module, panel }) => {
             Logger.ModuleManger.info(`${module}创建主面板: ${panel}`);
         });
 
-        SEASocketListener.on(CommandID.NOTE_USE_SKILL, 'receive', (data) => {
+        socketEmitter.on(CommandID.NOTE_USE_SKILL, 'receive', (data) => {
             const [fi, si] = data;
             Logger.BattleManager.info(`对局信息更新:
                 先手方:${fi.userId}
@@ -111,7 +111,7 @@ export default function App() {
                 使用技能: ${SkillXMLInfo.getName(si.skillId)}`);
         });
 
-        SEASocketListener.on(CommandID.USE_SKILL, 'send', (data) => {
+        socketEmitter.on(CommandID.USE_SKILL, 'send', (data) => {
             const [skillId] = data as [number];
             Logger.BattleManager.info(
                 `${FighterModelFactory.playerMode?.info.petName} 使用技能: ${SkillXMLInfo.getName(skillId)}`
@@ -129,7 +129,7 @@ export default function App() {
         const clean = () => {
             active = false;
             SEAModManager.teardown();
-            eventBus.unmount();
+            eventBus.dispose();
             document.body.removeEventListener('keydown', handleShortCut);
         };